requirements.txt
python-telegram-bot==20.3
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
cryptography==41.0.7
aiohttp==3.9.1
docker==6.1.3
python-dotenv==1.0.0
